pyproject.toml
[project]
name = "langgraph-v1-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.2.7",
    "jupyter>=1.1.1",
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langchain-community>=0.3.31",
    "langchain-experimental>=0.3.4",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-tavily>=0.2.12",
    "langchain-teddynote>=0.5.4",
    "langgraph>=1.0.1",
    "langgraph-supervisor>=0.0.31",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langgraph-checkpoint-redis>=0.1.2",
    "matplotlib>=3.10.7",
    "mcp>=1.21.0",
    "notebook>=7.4.7",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "psycopg2-binary",
    "psycopg[binary,pool]>=3.2.12",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "finance-datareader>=0.9.96",
    "yfinance>=0.2.66",
    "langchain-huggingface",
    "langchain-text-splitters",
    "faiss-cpu",
    "sentence-transformers",
    "pypdf",
    "weasyprint>=60.0",
    "markdown>=3.7",
]
